pyproject.toml
[project]
name = "guestflow-agent"
version = "0.1.0"
description = "Realtime Hotel Booking Voice Agent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "fastrtc>=0.2.0",
    "uvicorn>=0.30.6",
    "twilio>=9.3.2",
    "pydantic[email]>=2.9.2",
    "pydantic-settings>=2.5.2",
    "langgraph>=0.2.22",
    "langchain-core>=0.3.5",
    "openai>=1.47.1", # For vLLM OpenAI API compatibility
    "websockets>=13.1",
    "sqlalchemy>=2.0.35",
    "requests>=2.32.3",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.10",
    "superlinked>=12.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
